MAINT: Commented and removed unused code
